wip! cogl: provide way to specify framebuffer format
At the moment the onscreen framebuffer pixel format is
recorded as having the hardcoded value:

COGL_PIXEL_FORMAT_RGBA_8888_PRE

but there's nothing to actually enforce that format is getting
used.  In fact, mutter is using a gbm surface with format
GBM_FORMAT_ARGB8888 which translates to COGL_PIXEL_FORMAT_BGRA_8888_PRE.

This commit modifies the onscreen_init vfunc to allow the
implementation to pass back the format actually being used,
and changes mutter to do that.

This fixes readpixels color inversion.